+++ manifest/product_octez.ml
@@ -5676,6 +5676,7 @@ let octez_agnostic_baker_lib =
         octez_node_config;
         octez_client_commands |> open_;
         octez_profiler |> open_;
+        octez_stdlib_unix |> open_;
       ]
 
 (* PROTOCOL PACKAGES *)