+++ src/lib_agnostic_baker/daemon.ml
@@ -27,11 +27,32 @@ type state = {
   node_endpoint : string;
   mutable current_baker : baker option;
   mutable old_baker : baker_to_kill option;
+  keep_alive : bool;
 }
 
 type t = state
 
 (* ---- Baker Process Management ---- *)
+
+let rec retry_on_disconnection ~emit node_addr f =
+  let open Lwt_result_syntax in
+  let*! result = f () in
+  match result with
+  | Ok () -> return_unit
+  | Error (Lost_node_connection :: _ | Cannot_connect_to_node _ :: _) ->
+      let* _level =
+        Utils.retry
+          ~emit
+          ~max_delay:10.
+          ~delay:1.
+          ~factor:1.5
+          ~tries:max_int
+          ~is_error:(function Cannot_connect_to_node _ -> true | _ -> false)
+          (fun node_addr -> Rpc_services.get_level ~node_addr)
+          node_addr
+      in
+      retry_on_disconnection ~emit node_addr f
+  | Error trace -> fail trace
 
 (** [run_thread ~protocol_hash ~baker_commands ~cancel_promise]
     returns the main running thread for the baker given its protocol [~protocol_hash],
@@ -277,8 +298,8 @@ let may_start_initial_baker state =
   in
   may_start ~head_stream:None ()
 
-let create ~node_endpoint =
-  {node_endpoint; current_baker = None; old_baker = None}
+let create ~node_endpoint ~keep_alive =
+  {node_endpoint; current_baker = None; old_baker = None; keep_alive}
 
 let run state =
   let open Lwt_result_syntax in
@@ -289,8 +310,27 @@ let run state =
         let*! () = Events.(emit stopping_daemon) () in
         Lwt.return_unit)
   in
-  let* () = may_start_initial_baker state in
-  let* head_stream = monitor_heads ~node_addr in
+  let* () =
+    if state.keep_alive then
+      retry_on_disconnection
+        ~emit:(Events.emit Agnostic_baker_events.cannot_connect)
+        node_addr
+        (fun () -> may_start_initial_baker state)
+    else may_start_initial_baker state
+  in
+  let monitor_voting_periods () =
+    let* head_stream = monitor_heads ~node_addr in
+    monitor_voting_periods ~state head_stream
+  in
   (* Monitoring voting periods through heads monitoring to avoid
      missing UAUs. *)
-  Lwt.pick [monitor_voting_periods ~state head_stream; baker_thread ~state]
+  Lwt.pick
+    [
+      (* We do not care if --keep-alive is provided, if the baker thread doesn't
+         have the argument it'll abort the process anyway. *)
+      retry_on_disconnection
+        ~emit:(fun _ -> Lwt.return_unit)
+        node_addr
+        monitor_voting_periods;
+      baker_thread ~state;
+    ]