+++ CHANGES.rst
@@ -87,6 +87,8 @@ Baker
 Agnostic Baker
 --------------
 
+- Fix the support of ``--keep-alive`` for the agnostic baker. (MR :gl:`!17685`)
+
 - The agnostic baker binary becomes ``octez-agnostic-baker``. (MR :gl:`!17491`)
 
 - The agnostic baker now has the same CLI as the classical baker, getting rid of the